requirements.txt
python-telegram-bot==20.7
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
reportlab==4.0.9
pillow==10.1.0